% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi_mdl.R
\name{gsi_mdl}
\alias{gsi_mdl}
\title{Run single baseline GSI model}
\usage{
gsi_mdl(
  dat_in,
  nreps,
  nburn,
  thin,
  nchains,
  nadapt = 0,
  keep_burn = FALSE,
  cond_gsi = TRUE,
  harvest = NULL,
  file = NULL,
  seed = NULL
)
}
\arguments{
\item{dat_in}{Input data (as a list object).}

\item{nreps}{Total number of iterations (includes burn-ins).}

\item{nburn}{Number of warm-up runs.}

\item{thin}{Frequency to thin the output.}

\item{nchains}{Number of independent MCMC processes.}

\item{nadapt}{Number of adaptation run (default is 0). Only available when
running model in fully Bayesian mode.}

\item{keep_burn}{Logical (default = FALSE). To save the burn-ins or not.}

\item{cond_gsi}{Logical (default = TRUE). To run the model in conditional GSI mode.}

\item{harvest}{An optional harvest number for calculating the probability of p = 0. A proportion is considered as 0 if it's less than 5e-7 by default. If harvest number is provided, p = 0 is calculated as 0.5 / harvest of that stock.}

\item{file}{File path to save the output in RDS file. Need to type out the full path and extension \code{.Rds}. Default = NULL for not saving the output.}

\item{seed}{Random seed for reproducibility. Default = NULL (no random seed).}
}
\value{
A list containing:
\itemize{
\item Summary of the estimates
\item Trace of posterior samples
\item Individual assignment history
}
}
\description{
Run single baseline GSI model
}
\examples{
# prep input data
gsi_data <- prep_gsi_data(mixture_data = mix, baseline_data = baseline, pop_info = pops211)

# run model
gsi_out <- gsi_mdl(gsi_data, 10, 5, 1, 4)

}
