% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harv_summ.R
\name{harv_func}
\alias{harv_func}
\title{Sample total harvest numbers from a lognormal distribution}
\usage{
harv_func(x, n = 5000, seed = NULL, ...)
}
\arguments{
\item{x}{A vector consists of harvest mean and CV.}

\item{n}{Number of sample draws.}

\item{seed}{Optional random seed.}
}
\value{
A vector of harvest numbers drawn from a lognormal distribution with specified mean and CV.
}
\description{
Sample total harvest numbers from a lognormal distribution
}
\examples{
tot_harv <- harv_func(c(500, 0.05))

}
