% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harv_summ.R
\name{harv_summ}
\alias{harv_summ}
\title{Summarize trace output for stock-specific harvest}
\usage{
harv_summ(mdl_out, mdl_dat)
}
\arguments{
\item{mdl_out}{Output of GSI model run}

\item{mdl_dat}{Input data for GSI model}
}
\value{
A tibble of harvest estimates for each reporting group as columns and MCMC iterations as rows.
}
\description{
Summarize trace output for stock-specific harvest
}
\examples{
# prep input data
gsi_data <- prep_gsi_data(mixture_data = mix, baseline_data = baseline, pop_info = pops211)

# run model
tot_harv <- harv_func(c(500, 0.05))
gsi_out <- gsi_mdl(gsi_data, 10, 5, 1, 1, harvest = tot_harv)

# summarize individual assignments
harv_trace <- harv_summ(gsi_out, gsi_data)

}
