% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_summ.R
\name{indiv_assign}
\alias{indiv_assign}
\title{Calculate individual assignment using likelihood}
\usage{
indiv_assign(gsi_dat)
}
\arguments{
\item{gsi_dat}{Input data for GSI model}
}
\value{
An array of assignment probabilities for each reporting group as columns and individual fish in the mixture as rows.
}
\description{
Calculate individual assignment using likelihood
}
\examples{
# prep input data
gsi_data <- prep_gsi_data(mixture_data = mix, baseline_data = base, pop_info = pops211)

assign_like <- indiv_assign(gsi_data)

}
