% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_summ.R
\name{indiv_summ}
\alias{indiv_summ}
\title{Summarize individual assignment history output}
\usage{
indiv_summ(mdl_out, mdl_dat)
}
\arguments{
\item{mdl_out}{Output of GSI model run}

\item{mdl_dat}{Input data for GSI model}
}
\value{
A tibble of assignment probabilities for each reporting group as columns and individual fish in the mixture as rows.
}
\description{
Summarize individual assignment history output
}
\examples{
# prep input data
gsi_data <- prep_gsi_data(mixture_data = mix, baseline_data = baseline, pop_info = pops211)

# run model
gsi_out <- gsi_mdl(gsi_data, 10, 5, 1, 1)

# summarize individual assignments
indiv_ass <- indiv_summ(gsi_out, gsi_data)

}
