% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_summ.R
\name{rate_calc}
\alias{rate_calc}
\title{Calculate error and accuracy rates for plotting ROC}
\usage{
rate_calc(scaled_like, threshold, unit)
}
\arguments{
\item{threshold}{Cut-off for accepting individual assignment.}

\item{unit}{Which reporting unit to calculate.}

\item{like_dat}{Output from self_assign_le.}
}
\value{
An array with error/accuracy rates for plotting ROC.
}
\description{
Calculate error and accuracy rates for plotting ROC
}
\examples{
\donotrun{
rates <- rate_calc(scaled_like, threshold = 0.5, unit = "Russia")
}

}
