% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_summ.R
\name{self_assign_le}
\alias{self_assign_le}
\title{Self-assignment using baseline data}
\usage{
self_assign_le(base_dat, pop_info)
}
\arguments{
\item{base_dat}{Baseline data set.}

\item{pop_info}{Population information for the baseline.}
}
\value{
An array of assignment probabilities for each reporting group as columns and individual fish in the baseline as rows. Each individual has a known identity.
}
\description{
Self-assignment using baseline data
}
\examples{
\dontrun{
scaled_like <- self_assign_le(baseline, pops211)
}

}
