% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{tr_plot}
\alias{tr_plot}
\title{Plot MCMC trace}
\usage{
tr_plot(mdl_out, trace_obj, pop_info = NULL)
}
\arguments{
\item{mdl_out}{Model output object name.}

\item{trace_obj}{Trace from the model output.}

\item{pop_info}{Population information. A tibble with columns
collection (collection names), repunit (reporting unit names),
and grpvec (group numbers).}
}
\value{
Trace plot in ggplot
}
\description{
Plot MCMC trace
}
\examples{
# set up input data
gsi_data <- prep_gsi_data(mixture_data = mix, baseline_data = baseline, pop_info = pops211)

# run model
gsi_out <- gsi_mdl(gsi_data, 10, 5, 1, 1)

# trace plot
tr_plot(mdl_out = gsi_out, trace_obj = "trace", pop_info = gsi_out$groups)
}
