% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi_mdl.R
\name{gsi_mdl}
\alias{gsi_mdl}
\title{Run single baseline GSI model}
\usage{
gsi_mdl(
  dat_in,
  nreps,
  nburn,
  thin,
  nchains,
  nadapt = 0,
  keep_burn = FALSE,
  cond_gsi = TRUE,
  file = NULL,
  seed = NULL
)
}
\arguments{
\item{dat_in}{Input data (as a list object).}

\item{nreps}{Total number of iterations (includes burn-ins).}

\item{nburn}{Number of warm-up runs.}

\item{thin}{Frequency to thin the output.}

\item{nchains}{Number of independent MCMC processes.}

\item{nadapt}{Number of adaptation run (default is 0). Only available when
running model in fully Bayesian mode.}

\item{keep_burn}{Logical (default = FALSE). To save the burn-ins or not.}

\item{cond_gsi}{Logical (default = TRUE). To run the model in conditional GSI mode.}

\item{file}{File path to save the output in RDS file. Need to type out the full path and extension \code{.Rds}. Default = NULL for not saving the output.}

\item{seed}{Random seed for reproducibility. Default = NULL (no random seed).}
}
\value{
A list containing:
\itemize{
\item Summary of the estimates
\item Trace of posterior samples
\item Individual assignment history
}
}
\description{
Run single baseline GSI model
}
\examples{
\dontrun{
gsi_out <- gsi_mdl(gsi_data, 10, 5, 1, 4)
}

}
