% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{tr_plot}
\alias{tr_plot}
\title{Plot MCMC trace}
\usage{
tr_plot(obj, nburn = 0, thin = 1, name_order = NULL)
}
\arguments{
\item{obj}{Trace from the model output.}

\item{nburn}{Number of burn-in you set up when you ran the model.
Default is 0 if you didn't save the burn-ins (keep_burn = FALSE).}

\item{thin}{Number of thinning you set up when you ran the model.
Default is 1 (no thinning).}

\item{name_order}{Arrange the reporting groups as you wish. Leave it empty
if you want to accept the default.}
}
\value{
Trace plot in ggplot
}
\description{
Plot MCMC trace
}
\examples{
# set up input data
gsi_data <- prep_gsi_data(mixture_data = mix, baseline_data = baseline, pop_info = pops211)

# run model
gsi_out <- gsi_mdl(gsi_data, 10, 5, 1, 1)

# trace plot
tr_plot(obj = gsi_out$trace)
}
