% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{tr_plot}
\alias{tr_plot}
\title{Plot MCMC trace}
\usage{
tr_plot(obj, nburn = 0, thin = 1, name_order = NULL)
}
\arguments{
\item{obj}{Trace from the model output.}

\item{nburn}{Number of burn-in you set up when you ran the model.
Default is 0 if you didn't save the burn-ins (keep_burn = FALSE).}

\item{thin}{Number of thinning you set up when you ran the model.
Default is 1 (no thinning).}

\item{name_order}{Arrange the reporting groups as you wish. Leave it empty
if you want to accept the default.}
}
\value{
Trace plot in ggplot
}
\description{
Plot MCMC trace
}
\examples{
# set up input data and run multistage model
msgsi_dat <-
  prep_msgsi_data(mixture_data = mix,
  baseline1_data = base_templin, baseline2_data = base_yukon,
  pop1_info = templin_pops211, pop2_info = yukon_pops50, sub_group = 3:5)

msgsi_out <- msgsi_mdl(msgsi_dat, nreps = 25, nburn = 15, thin = 1, nchains = 1)

# trace plot
tr_plot(obj = msgsi_out$trace_comb)
}
